% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2d_concentration.R
\name{h2d_concentration}
\alias{h2d_concentration}
\title{Import concentration data from HYDRUS 2D output}
\usage{
h2d_concentration(path)
}
\arguments{
\item{path}{Path to HYDRUS 2D project containing 'MESHTRIA.TXT' and
'CONC<n>.TXT' with <n> being 1, 2, ...,}
}
\value{
Returns a tibble with 5 columns.
  'timestep': 'Print Times' in units defined in HYDRUS 'Time Information',
  'x': x-coordinate of mesh node,
  'y': y-coordinate of mesh node,
  'parameter': 'concentrationX' for concentration (concentration1,
    concentration2, ..., concentration7),
  'value': numerical value in units defined in HYDRUS [-].
}
\description{
This function imports concentration data from a simulation successfully run
in HYDRUS 2D. Use 'h3d_' functions to import HYDRUS 3D results. Simulation
results ('Mesh Information' and 'Concentrations') have to be exported to
ASCII prior to running this function using the HYDRUS 2D/3D GUI
('Results' --> 'Convert Output to ASCII').
}
\examples{
  h2d_concentration(path = "data")
}
\references{
https://www.pc-progress.com/downloads/Pgm_Hydrus3D2/HYDRUS3D%20User%20Manual.pdf
}
\seealso{
Other HYDRUS 2D functions: \code{\link{h2d_edges}},
  \code{\link{h2d_head}}, \code{\link{h2d_nodes}},
  \code{\link{h2d_temperature}},
  \code{\link{h2d_velocity}}, \code{\link{h2d_vwc}}
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
