% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2d_edges.R
\name{h2d_edges}
\alias{h2d_edges}
\title{Import edge coordinates from HYDRUS 2D mesh}
\usage{
h2d_edges(path)
}
\arguments{
\item{path}{Path to HYDRUS 2D project containing 'MESHTRIA.TXT'.}
}
\value{
Returns a tibble with 7 columns.
  'edgeID': edge number,
  'startNode': number of the beginning node of the edge,
  'endNode': number of the ending node of the edge,
  'loppNode': number of the left opposite node,
  'roppNode': number of the right opposite node,
  'lTriangle': number of the left triangle,
  'rTriangle': number of the right triangle.
  Note: 'roppNode', 'lTriangle', and 'lTriangle' are not always defined
}
\description{
This function imports edge coordinates from a simulation successfully run in
HYDRUS 2D. Use 'h3d_' functions to import HYDRUS 3D results. Simulation
results (including mesh information) have to be exported to ASCII prior to
running this function using the HYDRUS 2D/3D GUI ('Results' -->
'Convert Output to ASCII').
}
\examples{
  h2d_edges(path = "data")
}
\references{
https://www.pc-progress.com/downloads/Pgm_Hydrus3D2/HYDRUS3D%20User%20Manual.pdf
}
\seealso{
Other HYDRUS 2D functions: \code{\link{h2d_concentration}},
  \code{\link{h2d_head}}, \code{\link{h2d_nodes}},
  \code{\link{h2d_temperature}},
  \code{\link{h2d_velocity}}, \code{\link{h2d_vwc}}
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
