% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2d_head.R
\name{h2d_head}
\alias{h2d_head}
\title{Import pressure head data from HYDRUS output}
\usage{
h2d_head(path)
}
\arguments{
\item{path}{Path to HYDRUS 2D project containing 'MESHTRIA.TXT' and 'H.TXT'.}
}
\value{
Returns a tibble with 5 columns.
  'timestep': 'Print Times' in units defined in HYDRUS 'Time Information'.
  'x': x-coordinate of HYDRUS mesh node.
  'y': y-coordinate of HYDRUS mesh node.
  'parameter': 'head' for pressure head.
  'value': numerical value in units defined in HYDRUS ([L] for pressure head)
}
\description{
This function imports pressure head data from a simulation successfully run in HYDRUS 2D.
Use 'h3d_' functions to import HYDRUS 3D results.
Simulation results ('Mesh Information' and 'Pressure Heads') have to be exported to ASCII
prior to running this function using the HYDRUS 2D/3D GUI ('Results' --> 'Convert Output to ASCII').
}
\references{
https://www.pc-progress.com/downloads/Pgm_Hydrus3D2/HYDRUS3D%20User%20Manual.pdf
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
\keyword{IO}
