% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2d_nodes.R
\name{h2d_nodes}
\alias{h2d_nodes}
\title{Import node coordinates from HYDRUS mesh}
\usage{
h2d_nodes(path)
}
\arguments{
\item{path}{Path to HYDRUS 2D/3D project containing 'MESHTRIA.TXT'.}
}
\value{
Returns a tibble with 3 columns: 'nodeID', 'x', 'y' with 'x' and 'y'
  being the cartesian coordinates of node 'nodeID'.
}
\description{
This function imports node coordinates from a simulation successfully run in HYDRUS 2D.
Use 'h3d_' functions to import HYDRUS 3D results.
Simulation results (including mesh information) have to be exported to ASCII
prior to running this function using the HYDRUS 2D/3D GUI ('Results' --> 'Convert Output to ASCII').
}
\examples{
  import_nodes(path = "C://HYDRUS_Project/Project_Name")
}
\references{
https://www.pc-progress.com/downloads/Pgm_Hydrus3D2/HYDRUS3D%20User%20Manual.pdf
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
\keyword{IO}
