% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2d_velocity.R
\name{h2d_velocity}
\alias{h2d_velocity}
\title{Import velocity data from HYDRUS output}
\usage{
h2d_velocity(path)
}
\arguments{
\item{path}{Path to HYDRUS 2D project containing 'MESHTRIA.TXT' and 'V.TXT'.}
}
\value{
Returns a tibble with 6 columns.
  'timestep': 'Print Times' in units defined in HYDRUS 'Time Information'.
  'x': x-coordinate of HYDRUS mesh node.
  'y': y-coordinate of HYDRUS mesh node.
  'parameter': 'velocity' for velocities.
  'value': numerical value for magnitude of velocity vector in units defined in HYDRUS [L T^{-1}]
  'direction' numerical value for direction of velocity vector
}
\description{
This function imports velocity data from a simulation successfully run in HYDRUS 2D.
Use 'h3d_' functions to import HYDRUS 3D results.
Simulation results ('Mesh Information' and 'Velocities') have to be exported to ASCII
prior to running this function using the HYDRUS 2D/3D GUI ('Results' --> 'Convert Output to ASCII').
}
\references{
https://www.pc-progress.com/downloads/Pgm_Hydrus3D2/HYDRUS3D%20User%20Manual.pdf
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
\keyword{IO}
