% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vwc.R
\name{import_vwc}
\alias{import_vwc}
\title{Import water content data from HYDRUS output}
\usage{
import_vwc(path)
}
\arguments{
\item{path}{Path to HYDRUS 2D/3D project containing 'MESHTRIA.TXT' and 'TH.TXT'.}
}
\value{
Returns a tibble with 5 columns.
  'timestep': 'Print Times' in units defined in HYDRUS 'Time Information'.
  'x': x-coordinate of HYDRUS mesh node.
  'y': y-coordinate of HYDRUS mesh node.
  'parameter': 'vwc' for volumetric water content.
  'value': numerical value in units defined in HYDRUS (dimensionless for VWC)
}
\description{
This function imports volumentric water content (VWC) data from a simulation successfully run in HYDRUS 2D/3D.
Simulation results ('Mesh Information' and 'Water Contents') have to be exported to ASCII
prior to running this function using the HYDRUS 2D/3D GUI ('Results' --> 'Convert Output to ASCII').
}
\examples{
  import_vwc(path = "C://HYDRUS_Project/Project_Name")
}
\references{
https://www.pc-progress.com/downloads/Pgm_Hydrus3D2/HYDRUS3D%20User%20Manual.pdf
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
\keyword{IO}
