% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2d_head.R
\name{h2d_head}
\alias{h2d_head}
\title{Import pressure head data from HYDRUS 2D output}
\usage{
h2d_head(path)
}
\arguments{
\item{path}{Path to HYDRUS 2D project containing 'MESHTRIA.TXT' and 'H.TXT'.}
}
\value{
Returns a tibble with 5 columns.
  'timestep': 'Print Times' in units defined in HYDRUS 'Time Information',
  'x': x-coordinate of mesh node,
  'y': y-coordinate of mesh node,
  'parameter': 'head' for pressure head,
  'value': numerical value in units defined in HYDRUS [L].
}
\description{
This function imports pressure head data from a simulation successfully run
in HYDRUS 2D. Use 'h3d_' functions to import HYDRUS 3D results. Simulation
results ('Mesh Information' and 'Pressure Heads') have to be exported to
ASCII prior to running this function using the HYDRUS 2D/3D GUI ('Results'
--> 'Convert Output to ASCII').
}
\examples{
h2d_head(path = "data")

\dontrun{
h2d_head(path = "H.TXT")
h2d_head(path = "Project/H.TXT")
}
}
\references{
https://www.pc-progress.com/downloads/Pgm_Hydrus3D2/HYDRUS3D%20User%20Manual.pdf
}
\seealso{
Other HYDRUS 2D functions: \code{\link{h2d_concentration}},
  \code{\link{h2d_edges}}, \code{\link{h2d_nodes}},
  \code{\link{h2d_temperature}},
  \code{\link{h2d_velocity}}, \code{\link{h2d_vwc}}
}
\author{
Jan Knappe, \email{jan.knappe@gmail.com}
}
